Extract constants to separate module
Removes hardcoded stage name of dev throughout the code and
consolidates to a single location.
