+++ chalice/constants.py
@@ -1,0 +1,72 @@
+
+# This is the version that's written to the config file
+# on a `chalice new-project`.  It's also how chalice is able
+# to know when to warn you when changing behavior is introduced.
+CONFIG_VERSION = '2.0'
+
+
+TEMPLATE_APP = """\
+from chalice import Chalice
+
+app = Chalice(app_name='%s')
+
+
+@app.route('/')
+def index():
+    return {'hello': 'world'}
+
+
+# The view function above will return {"hello": "world"}
+# whenever you make an HTTP GET request to '/'.
+#
+# Here are a few more examples:
+#
+# @app.route('/hello/{name}')
+# def hello_name(name):
+#    # '/hello/james' -> {"hello": "james"}
+#    return {'hello': name}
+#
+# @app.route('/users', methods=['POST'])
+# def create_user():
+#     # This is the JSON body the user sent in their POST request.
+#     user_as_json = app.json_body
+#     # Suppose we had some 'db' object that we used to
+#     # read/write from our database.
+#     # user_id = db.create_user(user_as_json)
+#     return {'user_id': user_id}
+#
+# See the README documentation for more examples.
+#
+"""
+
+
+GITIGNORE = """\
+.chalice/deployments/
+.chalice/venv/
+"""
+
+DEFAULT_STAGE_NAME = 'dev'
+
+
+LAMBDA_TRUST_POLICY = {
+    "Version": "2012-10-17",
+    "Statement": [{
+        "Sid": "",
+        "Effect": "Allow",
+        "Principal": {
+            "Service": "lambda.amazonaws.com"
+        },
+        "Action": "sts:AssumeRole"
+    }]
+}
+
+
+CLOUDWATCH_LOGS = {
+    "Effect": "Allow",
+    "Action": [
+        "logs:CreateLogGroup",
+        "logs:CreateLogStream",
+        "logs:PutLogEvents"
+    ],
+    "Resource": "arn:aws:logs:*:*:*"
+}
